requirements.txt
blinker==1.6.3
boto3==1.28.62
botocore==1.31.62
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.7.0
six==1.16.0
urllib3==1.26.17
Werkzeug==3.0.0
zipp==3.17.0
